#515 move leuchten form to new component
